server/requirements.txt
alembic==1.4.3
Babel==2.8.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
enum34==1.1.10
eventlet==0.28.0
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Cors==3.0.9
Flask-HTTPAuth==4.1.0
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gevent==20.9.0
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
monotonic==1.5
passlib==1.7.2
phonenumbers==8.12.10
pycparser==2.20
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-cache==0.5.2
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.19
twilio==6.45.4
urllib3==1.25.10
uWSGI==2.0.17.1
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.1.2


